requirements.txt
astroid==2.4.2
autopep8==1.5.5
click==7.1.2
colorama==0.4.4
fastapi==0.63.0
flake8==3.8.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
peewee==3.14.1
peewee-migrate==1.4.0
psycopg2==2.8.6
pycodestyle==2.6.0
pydantic==1.7.3
pyflakes==2.2.0
pylint==2.6.0
PyYAML==5.4.1
six==1.15.0
starlette==0.13.6
toml==0.10.2
wrapt==1.12.1
